core/refactor: split main_loop into modules, simplify, rename
fully AI generated.

prompts used:

1. simplification - ran this about 5 times
```
 make @core/translate/main_loop.rs simpler without changing semantics. if you cannot make it any simpler, just output: NO
  SIMPLIFICATIONS AVAILABLE. Focus on larger structural refactors and wins over small touchups and tidiness.
  The only success gate is: "make -C testing/runner run-rust" passes. no need for cargo build, cargo clippy, cargo fmt
  etc.
```

2. splitting

```
split main_loop.rs into separate submodules (perhaps in a subdirectory) without changing semantics. focus on sharing similar concerns per file and orchestrating from parent module.
  The only success gate is: "make -C testing/runner run-rust" passes. no need for cargo build, cargo clippy, cargo fmt
  etc.
```

3. simplification again a few times after the split

4. renaming:

```
read the modules in @core/translate/main_loop and rename any functions, structs or enums that do not accurately describe what they do. for example, if some function only handles hash join related logic but has a general name, it should explicitly mention hash joins. do this universally, not just for hash joins.
```
